backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
anthropic
python-dotenv
python-multipart
supabase
httpx
pytest
slowapi
ruff
